Update Main Page UI

+++ api/views/file.py
@@ -1,8 +1,10 @@
 from django.views import View
 from django.http import JsonResponse
-from app01.models import Avatars, Cover
+from app01.models import Avatars, Cover, UserInfo
 from django.core.files.uploadedfile import InMemoryUploadedFile
 from app01.models import avatar_delete, cover_delete
+from django.core.handlers.wsgi import WSGIRequest
+from django.db.models import Q
 
 class AvatarView(View):
     def post(self, request):
@@ -49,7 +51,20 @@ class AvatarView(View):
 
         # 使用models.py中定义好的删除头像方法
         avatar_query = Avatars.objects.filter(nid=nid)
-        avatar_delete(avatar_query.first())
+        if not avatar_query:
+            res['msg'] = 'Avatar has been deleted!'
+            return JsonResponse(res)
+
+        # 判断图片是否有人在使用
+        obj: Avatars = avatar_query.first()
+
+        userquery = UserInfo.objects.filter(Q(sign_status=1) | Q(sign_status=2))
+        for user in userquery:
+            if obj.url.url == user.avatar_url:
+                res['msg'] = 'This avatar is being used by user.'
+                return JsonResponse(res)
+
+        avatar_delete(obj)
         avatar_query.delete()
 
         res['code'] = 0
@@ -99,6 +114,9 @@ class CoverView(View):
 
             # 使用models.py中定义好的删除头像方法
             cover_query = Cover.objects.filter(nid=nid)
+            if not cover_query:
+                res['msg'] = 'Cover has been deleted!'
+                return JsonResponse(res)
             cover_delete(cover_query.first())
             cover_query.delete()
